requirements.txt
aniso8601==6.0.0
attrs==19.1.0
Click==7.0
Flask==1.0.3
Flask-JWT-Extended==3.24.1
flask-restplus==0.12.1
Flask-SQLAlchemy==2.3.2
gunicorn==19.3.0
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==3.0.1
MarkupSafe==1.1.1
mysql-connector-python==8.0.21
PyJWT==1.7.1
pyrsistent==0.15.2
pytz==2019.1
six==1.12.0
SQLAlchemy==1.3.18
Werkzeug==0.15.4
flask-swagger-ui==3.25.0
